+++ seqr/models.py
@@ -1182,7 +1182,7 @@ class RnaSeqSpliceOutlier(DeletableSampleMetadataModel):
                        'delta_intron_jaccard_index', 'mean_counts', 'total_counts', 'mean_total_counts']
 
 
-class PhenotypePrioritization(BulkOperationBase):
+class PhenotypePrioritization(ModelWithGUID):
     PARENT_FIELD = 'individual'
 
     individual = models.ForeignKey('Individual', on_delete=models.CASCADE, db_index=True)
@@ -1196,6 +1196,9 @@ class PhenotypePrioritization(BulkOperationBase):
 
     def __unicode__(self):
         return "%s:%s:%s" % (self.individual.individual_id, self.gene_id, self.disease_id)
+
+    def _compute_guid(self):
+        return 'PP%07d_%s' % (self.id, _slugify(str(self)))
 
     class Meta:
         json_fields = ['gene_id', 'tool', 'rank', 'disease_id', 'disease_name', 'scores']